refact (bac/orden) ordenamiento
