--- src/controllers/hotel/getHotelsByCity.js
@@ -1,12 +1,0 @@
-const Hotel = require("../../models/hotelModel");
-
-const getHotelsByCity = async (city) => {
-    const hotelsFound = await Hotel.find({ city: city });
-    if (hotelsFound.length) {
-        return hotelsFound;
-    } else {
-        throw new Error("No results.");
-    };
-};
-
-module.exports = getHotelsByCity;
