+++ src/controllers/trending/getTrendingHotels.js
@@ -1,9 +1,9 @@
 const TrendingHotels = require("../../models/trendingHotelsModel");
 
-const getTrendigHotels = async () => {
+const getTrendingHotels = async () => {
     const trendingHotels = await TrendingHotels.find();
     if (trendingHotels.length) return trendingHotels;
     throw new Error("No results.");
 };
 
-module.exports = getTrendigHotels;
+module.exports = getTrendingHotels;
