+++ src/handlers/hotel/hotelsByStateGet.js
@@ -1,9 +1,9 @@
-const { getHotelsByCity } = require("../controllers/hotel");
+const { getHotelsByCity } = require("../../controllers/hotel");
 
 const hotelsByStateGet = async (req, res) => {
     try {
-        const { city } = req.params;
-        const hotelsFound = await getHotelsByCity(city);
+        const { state } = req.params;
+        const hotelsFound = await getHotelsByCity(state);
         res.status(200).json(hotelsFound);
     } catch (error) {
         res.status(404).json({ error: error.message });
