+++ src/handlers/trending/trendingHotelsGet.js
@@ -1,8 +1,8 @@
-const { getTrendigHotels } = require("../controllers/trending");
+const { getTrendingHotels } = require("../../controllers/trending");
 
 const trendingHotelsGet = async (req, res) => {
     try {
-        const trendingResults = await getTrendigHotels();
+        const trendingResults = await getTrendingHotels();
         res.status(200).json(trendingResults);
     } catch (error) {
         res.status(404).json({ error: error.message });
