+++ src/handlers/user/getUserByEmail.js
@@ -1,10 +1,10 @@
-const User = require("../models/userModel");
+const User = require("../../models/userModel");
 
 const getUserByEmail = async (req, res) => {
     try {
-        const { email } = req.body;        
+        const { email } = req.body;
 
-        const user = await User.find({email});
+        const user = await User.find({ email });
         res.status(200).json(user)
 
     } catch (error) {
