+++ src/routes/destinationRoutes.js
@@ -1,6 +1,6 @@
 const { Router } = require('express');
 const countriesGet = require("../handlers/countriesGet.js");
-const stateGet = require("../handlers/stateGet.js");
+const stateGet = require("../handlers/user/stateGet.js");
 
 const destinationRoutes = Router();
