+++ src/routes/userRoutes.js
@@ -2,11 +2,11 @@ const { Router } = require('express');
 const { validationResult } = require('express-validator');
 
 const validationUserPost = require("../middlewares/validationUserPost")
-const userPost = require("../handlers/userPost.js");
-const userGet = require("../handlers/userGet.js");
-const userGetID = require("../handlers/userGetID.js");
-const userDelete = require("../handlers/userDelete.js");
-const userPut = require("../handlers/userPut.js");
+const userPost = require("../handlers/user/userPost.js");
+const userGet = require("../handlers/user/userGet.js");
+const userGetID = require("../handlers/user/userGetID.js");
+const userDelete = require("../handlers/user/userDelete.js");
+const userPut = require("../handlers/user/userPut.js");
 
 const userRoutes = Router();
